Rename balance command to wallet -balance.
